Spectral response of sensors

A collection of response spectra for various broadband sensors used for measuring ultraviolet and visible radiation. Each spectrun in the collection contains two variables, wavelengths (nm) at either regular or irregular intervals and spectral responsiveness (in energy units). Spectral data are in most cases normalized to one at the wavelength of maximum energy responsivity. Absolute calibration values are given only for data from a publication which reports on mulstiple units of the same type.

sensors.mspct

Format

A response_mspct object containing a response_spct objects as named members.

Each member spectrum contains two variables, with responsivity in most cases in relative energy units:

w.length (nm)

s.e.response (r.u.)

Note

In addition to this object containing the spectral data, this package provides character vectors useful for subsetting spectra by supplier, type, color, etc.

See also

source_spct and generic_mspct

Examples

names(sensors.mspct)

#>  [1] "Berger_UV_Biometer" "BF5"                "flat_e"            
#>  [4] "flat_q"             "CUV_5"              "PQS1"              
#>  [7] "UVS_A"              "UVS_B"              "UVS_E"             
#> [10] "LI_190"             "LI_200"             "LI_210"            
#> [13] "TOCON_blue4"        "SG01D_A"            "SG01D_B"           
#> [16] "SG01D_C"            "SG01L"              "SKE510"            
#> [19] "SKL310"             "SKP210"             "SKP215"            
#> [22] "SKR110_FR"          "SKR110_R"           "SKS1110"           
#> [25] "SKU421"             "SKU421a"            "SKU430a"           
#> [28] "SKU440a"            "SL_501_high_UVA"    "SL_501_low_UVA"    
#> [31] "SL_501_typical"     "SM60"               "E1c"               
#> [34] "BW_20"