Sensors responsive to different wavebands

Names of datasets containing the wavelengths and tabulated values spectral emittance for broadband sensors from various suppliers.

uv_sensors

uvc_sensors

uvb_sensors

erythemal_sensors

uva_sensors

par_sensors

vis_sensors

photometric_sensors

shortwave_sensors

pyranometer_sensors

red_sensors

far_red_sensors

blue_sensors

multichannel_sensors

Format

A vector of character strings.

See also

sensors.mspct

Examples

uv_sensors

#>  [1] "SG01D_C"            "SG01D_B"            "SM60"              
#>  [4] "SKU430a"            "UVS_B"              "SG01D_A"           
#>  [7] "SKU421"             "SKU421a"            "UVS_A"             
#> [10] "UVS_E"              "E1c"                "SKU440a"           
#> [13] "SL_501_high_UVA"    "SL_501_low_UVA"     "SL_501_typical"    
#> [16] "BW_20"              "Berger_UV_Biometer" "SG01L"             
#> [19] "CUV_5"             

uvc_sensors

#> [1] "SG01D_C"

uvb_sensors

#> [1] "SG01D_B" "SM60"    "SKU430a" "UVS_B"  

uva_sensors

#> [1] "SG01D_A" "SKU421"  "SKU421a" "UVS_A"  

par_sensors

#> [1] "SKP215" "SKE510" "SKP210" "PQS1"   "LI_190" "BF5"   

vis_sensors

#> [1] "SKL310" "LI_210"

shortwave_sensors

#> [1] "SKS1110" "LI_200" 

red_sensors

#> [1] "SKR110_R"

far_red_sensors

#> [1] "SKR110_FR"

blue_sensors

#> [1] "TOCON_blue4"

multichannel_sensors

#> [1] "SKR110_R"  "SKR110_FR"

# select PAR sensors
sensors.mspct[par_sensors]

#> $SKP215
#> Object: response_spct [736 x 2]
#> Wavelength range 382.04143 to 750.07094 nm, step 0.5007204 nm 
#> Time unit 1s
#> 
#> # A tibble: 736 x 2
#>    w.length s.q.response
#>       <dbl>        <dbl>
#>  1      382         1.62
#>  2      383         1.67
#>  3      383         1.68
#>  4      384         1.76
#>  5      384         1.79
#>  6      385         1.84
#>  7      385         1.93
#>  8      386         1.95
#>  9      386         1.96
#> 10      387         2.02
#> # ... with 726 more rows
#> 
#> $SKE510
#> Object: response_spct [708 x 2]
#> Wavelength range 381.40741 to 735.58113 nm, step 0.5009529 nm 
#> Time unit 1s
#> 
#> # A tibble: 708 x 2
#>    w.length s.e.response
#>       <dbl>        <dbl>
#>  1      381         2.03
#>  2      382         2.05
#>  3      382         2.08
#>  4      383         2.14
#>  5      383         2.21
#>  6      384         2.23
#>  7      384         2.26
#>  8      385         2.31
#>  9      385         2.38
#> 10      386         2.41
#> # ... with 698 more rows
#> 
#> $SKP210
#> Object: response_spct [796 x 2]
#> Wavelength range 351.22185 to 749.24941 nm, step 0.5006636 nm 
#> Time unit 1s
#> 
#> # A tibble: 796 x 2
#>    w.length s.q.response
#>       <dbl>        <dbl>
#>  1      351        0.431
#>  2      352        0.528
#>  3      352        0.639
#>  4      353        0.758
#>  5      353        0.894
#>  6      354        1.01 
#>  7      354        1.12 
#>  8      355        1.22 
#>  9      355        1.33 
#> 10      356        1.47 
#> # ... with 786 more rows
#> 
#> $PQS1
#> Object: response_spct [652 x 2]
#> Wavelength range 391.431 to 717.608 nm, step 0.5010399 nm 
#> Time unit 1s
#> 
#> # A tibble: 652 x 2
#>    w.length s.e.response
#>       <dbl>        <dbl>
#>  1      391       0.0296
#>  2      392       0.0396
#>  3      392       0.0496
#>  4      393       0.0589
#>  5      393       0.0679
#>  6      394       0.0786
#>  7      394       0.0911
#>  8      395       0.104 
#>  9      395       0.116 
#> 10      396       0.129 
#> # ... with 642 more rows
#> 
#> $LI_190
#> Object: response_spct [755 x 2]
#> Wavelength range 365.614 to 742.99 nm, step 0.5004987 nm 
#> Time unit 1s
#> 
#> # A tibble: 755 x 2
#>    w.length s.q.response
#>       <dbl>        <dbl>
#>  1      366      0.00181
#>  2      366      0.00188
#>  3      367      0.00195
#>  4      367      0.00202
#>  5      368      0.00209
#>  6      368      0.00217
#>  7      369      0.00224
#>  8      369      0.00231
#>  9      370      0.00238
#> 10      370      0.00245
#> # ... with 745 more rows
#> 
#> $BF5
#> Object: response_spct [945 x 2]
#> Wavelength range 307.56022 to 780 nm, step 0.5004659 nm 
#> Time unit 1s
#> 
#> # A tibble: 945 x 2
#>    w.length s.e.response
#>       <dbl>        <dbl>
#>  1      308      0.00558
#>  2      308      0.00569
#>  3      309      0.00580
#>  4      309      0.00592
#>  5      310      0.00603
#>  6      310      0.00614
#>  7      311      0.00625
#>  8      311      0.00635
#>  9      312      0.00636
#> 10      312      0.00638
#> # ... with 935 more rows
#>